Added services for managing users, locations, and forecasts
