+++ OutbackWeatherTrackerApplication/src/main/java/cab302softwaredevelopment/outbackweathertrackerapplication/controllers/widgets/BaseWidgetController.java
@@ -7,6 +7,7 @@ import cab302softwaredevelopment.outbackweathertrackerapplication.database.model
 import cab302softwaredevelopment.outbackweathertrackerapplication.database.model.HourlyForecast;
 import cab302softwaredevelopment.outbackweathertrackerapplication.database.model.Location;
 import cab302softwaredevelopment.outbackweathertrackerapplication.models.DateData;
+import cab302softwaredevelopment.outbackweathertrackerapplication.models.WidgetInfo;
 import cab302softwaredevelopment.outbackweathertrackerapplication.utils.WidgetConfig;
 
 import java.time.LocalDate;
@@ -15,30 +16,18 @@ import java.time.ZonedDateTime;
 public abstract class BaseWidgetController implements IConfigurableWidget {
     Location location;
 
+    /**
+     * Method to apply widget configuration.
+     * This will be called after the widget is instantiated by the factory.
+     *
+     * @param config Configuration data for the widget.
+     */
     @Override
     public void applyConfig(WidgetConfig config) {
         long locationId = config.getLocationId();
         location = new LocationDAO.LocationQuery()
                 .whereId(locationId)
                 .getSingleResult();
-        updateWidget();
-    }
-
-    public HourlyForecast getLatestHourlyForecast() {
-        int nowEpoch = (int) DateData.getNearestHourEpoch(ZonedDateTime.now());
-        HourlyForecast currentForecast = new HourlyForecastDAO.HourlyForecastQuery()
-                .whereLocation(location)
-                .whereTimestamp(nowEpoch)
-                .getSingleResult();
-        return currentForecast;
-    }
-
-    public DailyForecast getTodayForecast() {
-        DateData now = new DateData(LocalDate.now());
-        DailyForecast todayForecast = new DailyForecastDAO.DailyForecastQuery()
-                .whereTimestamp((int) now.getDayStartEpoch())
-                .getSingleResult();
-        return todayForecast;
     }
 
     public abstract void updateWidget();
