+++ OutbackWeatherTrackerApplication/src/main/java/cab302softwaredevelopment/outbackweathertrackerapplication/controllers/widgets/SoilWidgetController.java
@@ -1,21 +1,17 @@
 package cab302softwaredevelopment.outbackweathertrackerapplication.controllers.widgets;
 
 import cab302softwaredevelopment.outbackweathertrackerapplication.database.model.HourlyForecast;
+import cab302softwaredevelopment.outbackweathertrackerapplication.services.ForecastService;
+import cab302softwaredevelopment.outbackweathertrackerapplication.services.LocationService;
 import javafx.fxml.FXML;
 import javafx.scene.text.Text;
 
 public class SoilWidgetController extends BaseWidgetController {
     @FXML
-    public Text txtET;
-    @FXML
-    public Text txtLocation;
-    @FXML
-    public Text txtTemperature;
-    @FXML
-    public Text txtMoisture;
+    public Text txtET, txtLocation, txtTemperature, txtMoisture;
 
     public void updateWidget() {
-        HourlyForecast currentForecast = getLatestHourlyForecast();
+        HourlyForecast currentForecast = ForecastService.getLatestHourlyForecast(location);
         if (currentForecast != null) {
             txtLocation.setText(currentForecast.getLocation().getName());
             txtMoisture.setText(currentForecast.getSoil_moisture_0_to_10cm() + "%");
