+++ OutbackWeatherTrackerApplication/src/main/java/cab302softwaredevelopment/outbackweathertrackerapplication/controllers/widgets/TempWidgetController.java
@@ -1,6 +1,8 @@
 package cab302softwaredevelopment.outbackweathertrackerapplication.controllers.widgets;
 
 import cab302softwaredevelopment.outbackweathertrackerapplication.database.model.HourlyForecast;
+import cab302softwaredevelopment.outbackweathertrackerapplication.services.ForecastService;
+import cab302softwaredevelopment.outbackweathertrackerapplication.services.LocationService;
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
 import javafx.scene.text.Text;
@@ -12,7 +14,7 @@ public class TempWidgetController extends BaseWidgetController {
     public Text txtCurrTemp, txtSoilTemp, txtFeelsTemp, txtSurfTemp;
 
     public void updateWidget() {
-        HourlyForecast currentForecast = getLatestHourlyForecast();
+        HourlyForecast currentForecast = ForecastService.getLatestHourlyForecast(location);
         if (currentForecast != null) {
             txtCurrTemp.setText(currentForecast.getTemperature_2m() + "°");
             txtSoilTemp.setText(currentForecast.getSoil_temperature_0_to_10cm() + "°" );
