+++ OutbackWeatherTrackerApplication/src/main/java/cab302softwaredevelopment/outbackweathertrackerapplication/controllers/widgets/WeatherWidget2Controller.java
@@ -1,6 +1,7 @@
 package cab302softwaredevelopment.outbackweathertrackerapplication.controllers.widgets;
 
 import cab302softwaredevelopment.outbackweathertrackerapplication.database.model.DailyForecast;
+import cab302softwaredevelopment.outbackweathertrackerapplication.services.ForecastService;
 import javafx.fxml.FXML;
 import javafx.scene.chart.LineChart;
 import javafx.scene.chart.XYChart;
@@ -15,15 +16,13 @@ public class WeatherWidget2Controller extends BaseWidgetController {
     public Text txtLocation;
 
     public void updateWidget() {
-        DailyForecast todaysForecast = getTodayForecast();
-        if (todaysForecast != null) {
-            txtLocation.setText(todaysForecast.getLocation().getName());
-            XYChart.Series<String,Number> chart = updateChart();
-            lgGraph.getData().add(chart);
-            txtMinMaxTemp.setText(
-                    "H: " + todaysForecast.getTemperature_2m_max() + "° " +
-                    "L: " + todaysForecast.getTemperature_2m_min() + "°");
-        }
+        DailyForecast todaysForecast = ForecastService.getTodayForecast(location);
+        txtLocation.setText(todaysForecast.getLocation().getName());
+        XYChart.Series<String,Number> chart = updateChart();
+        lgGraph.getData().add(chart);
+        txtMinMaxTemp.setText(
+                "H: " + todaysForecast.getTemperature_2m_max() + "° " +
+                "L: " + todaysForecast.getTemperature_2m_min() + "°");
     }
 
     public XYChart.Series updateChart(){
