+++ OutbackWeatherTrackerApplication/src/main/java/cab302softwaredevelopment/outbackweathertrackerapplication/controllers/widgets/WeatherWidget3Controller.java
@@ -2,6 +2,10 @@ package cab302softwaredevelopment.outbackweathertrackerapplication.controllers.w
 
 import cab302softwaredevelopment.outbackweathertrackerapplication.database.model.HourlyForecast;
 import javafx.fxml.FXML;
+import cab302softwaredevelopment.outbackweathertrackerapplication.database.model.HourlyForecast;
+import cab302softwaredevelopment.outbackweathertrackerapplication.services.ForecastService;
+import cab302softwaredevelopment.outbackweathertrackerapplication.services.LocationService;
+import javafx.fxml.FXML;
 import javafx.scene.image.ImageView;
 import javafx.scene.text.Text;
 
@@ -12,7 +16,25 @@ public class WeatherWidget3Controller extends BaseWidgetController {
     public Text txtTemperature, txtLocation, txtPrecipitation, txtWind, txtHumidity, txtFeelsLike;
 
     public void updateWidget() {
-        HourlyForecast currentForecast = getLatestHourlyForecast();
+        HourlyForecast currentForecast = ForecastService.getLatestHourlyForecast(location);
+
+        if (currentForecast != null) {
+            txtTemperature.setText(currentForecast.getTemperature_2m() + "°");
+            txtLocation.setText(currentForecast.getLocation().getName());
+            txtPrecipitation.setText(currentForecast.getPrecipitation() + "");
+            txtWind.setText(currentForecast.getWind_speed_10m() + "km/h");
+            txtHumidity.setText(currentForecast.getRelative_humidity_2m() + "%");
+            txtFeelsLike.setText(currentForecast.getApparent_temperature() + "°");
+        }
+    }
+    public void updateWeatherWidget(int temp, String location, String precipitation, int wind, int humid, int feels){
+        temperatureText.setText(temp + "°");
+        Location.setText(location);
+        Precipitation.setText(precipitation);
+        Wind.setText(wind + "km/h");
+        Humidity.setText(humid + "%");
+        FeelsLike.setText(feels + "°");
+// random comment
 
         if (currentForecast != null) {
             txtTemperature.setText(currentForecast.getTemperature_2m() + "°");
