+++ OutbackWeatherTrackerApplication/src/main/java/cab302softwaredevelopment/outbackweathertrackerapplication/controllers/widgets/WeatherWidget4Controller.java
@@ -1,6 +1,10 @@
 package cab302softwaredevelopment.outbackweathertrackerapplication.controllers.widgets;
 
 import cab302softwaredevelopment.outbackweathertrackerapplication.database.model.HourlyForecast;
+import cab302softwaredevelopment.outbackweathertrackerapplication.services.ForecastService;
+import cab302softwaredevelopment.outbackweathertrackerapplication.services.LocationService;
+import javafx.fxml.FXML;
+import cab302softwaredevelopment.outbackweathertrackerapplication.database.model.HourlyForecast;
 import javafx.fxml.FXML;
 import javafx.scene.image.ImageView;
 import javafx.scene.text.Text;
@@ -12,7 +16,7 @@ public class WeatherWidget4Controller extends BaseWidgetController {
     public ImageView imgSunrise, imgSunset;
 
     public void updateWidget() {
-        HourlyForecast currentForecast = getLatestHourlyForecast();
+        HourlyForecast currentForecast = ForecastService.getLatestHourlyForecast(location);
 
         if (currentForecast != null) {
             txtLocation.setText(currentForecast.getLocation().getName());
