+++ OutbackWeatherTrackerApplication/src/main/java/cab302softwaredevelopment/outbackweathertrackerapplication/controllers/widgets/WeatherWidgetController.java
@@ -1,5 +1,10 @@
 package cab302softwaredevelopment.outbackweathertrackerapplication.controllers.widgets;
 
+import cab302softwaredevelopment.outbackweathertrackerapplication.database.model.DailyForecast;
+import cab302softwaredevelopment.outbackweathertrackerapplication.database.model.HourlyForecast;
+import cab302softwaredevelopment.outbackweathertrackerapplication.models.WeatherCondition;
+import cab302softwaredevelopment.outbackweathertrackerapplication.services.ForecastService;
+import cab302softwaredevelopment.outbackweathertrackerapplication.services.LocationService;
 import cab302softwaredevelopment.outbackweathertrackerapplication.ApplicationEntry;
 import cab302softwaredevelopment.outbackweathertrackerapplication.database.model.DailyForecast;
 import cab302softwaredevelopment.outbackweathertrackerapplication.database.model.HourlyForecast;
@@ -18,9 +23,20 @@ public class WeatherWidgetController extends BaseWidgetController {
   private Text txtWeatherCondition, txtDateTime, txtTemperature, txtMinMaxTemp;
 
   public void updateWidget() {
-    HourlyForecast currentForecast = getLatestHourlyForecast();
-    DailyForecast todayForecast = getTodayForecast();
+    HourlyForecast currentForecast = ForecastService.getLatestHourlyForecast(location);
+    DailyForecast todayForecast = ForecastService.getTodayForecast(location);
 
+    if (currentForecast != null) {
+      WeatherCondition condition = WeatherCondition.fromReading(currentForecast);
+      txtWeatherCondition.setText(condition.getName());
+      Image weatherIcon = new Image(condition.getImagePath());
+      imgWeatherIcon.setImage(weatherIcon);
+      txtDateTime.setText("00:00");
+      txtTemperature.setText(currentForecast.getTemperature_2m() + "°");
+      txtMinMaxTemp.setText(
+              "H: " + todayForecast.getTemperature_2m_max() + "° " +
+              "L: " + todayForecast.getTemperature_2m_min() + "°");
+    }
     if (currentForecast != null && todayForecast != null) {
       WeatherCondition condition = WeatherCondition.fromReading(currentForecast);
       txtWeatherCondition.setText(condition.getName());
