+++ OutbackWeatherTrackerApplication/src/main/java/cab302softwaredevelopment/outbackweathertrackerapplication/controllers/widgets/WindWidgetController.java
@@ -1,6 +1,9 @@
 package cab302softwaredevelopment.outbackweathertrackerapplication.controllers.widgets;
 
 import cab302softwaredevelopment.outbackweathertrackerapplication.database.model.HourlyForecast;
+import cab302softwaredevelopment.outbackweathertrackerapplication.services.ForecastService;
+import cab302softwaredevelopment.outbackweathertrackerapplication.services.LocationService;
+import javafx.fxml.FXML;
 import javafx.fxml.FXML;
 import javafx.scene.text.Text;
 import javafx.scene.control.Label;
@@ -12,7 +15,7 @@ public class WindWidgetController extends BaseWidgetController {
     public Text txtLocation, txtWindspeed, txtGustSpeed, txtNSEW;
 
     public void updateWidget() {
-        HourlyForecast currentForecast = getLatestHourlyForecast();
+        HourlyForecast currentForecast = ForecastService.getLatestHourlyForecast(location);
         if (currentForecast == null) {
             // Display error?
         } else {
