+++ OutbackWeatherTrackerApplication/src/main/java/cab302softwaredevelopment/outbackweathertrackerapplication/services/LocationService.java
@@ -1,0 +1,18 @@
+package cab302softwaredevelopment.outbackweathertrackerapplication.services;
+
+import cab302softwaredevelopment.outbackweathertrackerapplication.database.dao.DailyForecastDAO;
+import cab302softwaredevelopment.outbackweathertrackerapplication.database.dao.HourlyForecastDAO;
+import cab302softwaredevelopment.outbackweathertrackerapplication.database.dao.LocationDAO;
+import cab302softwaredevelopment.outbackweathertrackerapplication.database.model.DailyForecast;
+import cab302softwaredevelopment.outbackweathertrackerapplication.database.model.HourlyForecast;
+import cab302softwaredevelopment.outbackweathertrackerapplication.database.model.Location;
+import cab302softwaredevelopment.outbackweathertrackerapplication.models.DateData;
+
+import java.time.LocalDate;
+import java.time.ZonedDateTime;
+
+public class LocationService {
+    private static LocationDAO locationDAO = new LocationDAO();
+
+
+}
